+++ src/main/java/com/epam/esm/service/CertificateService.java
@@ -20,5 +20,5 @@ public interface CertificateService extends CommonService<CertificateDTO, Long> 
 
     BigInteger getCountOfCertificate();
 
-    int getCountOfCertificateWithQuery(ParametersSpecificationDTO parameters);
+    Long getCountOfCertificateWithQuery(ParametersSpecificationDTO parameters);
 }
