+++ src/main/java/com/epam/esm/service/impl/CertificateServiceImpl.java
@@ -36,7 +36,7 @@ public class CertificateServiceImpl implements CertificateService {
     private final MapperDTO mapperDTO;
 
     @Override
-    public int getCountOfCertificateWithQuery(ParametersSpecificationDTO parameters) {
+    public Long getCountOfCertificateWithQuery(ParametersSpecificationDTO parameters) {
         return certificateDAO.getCountOfCertificateWithQuery(parameters);
     }
